tool-tem-v01/requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.5
pillow==10.4.0
python-barcode==0.15.1
